Use native D-Bus and libnotify implementation
This change makes the libnotify-compatible notifications more resilient
by switching to a different library that uses a native implementation of
the D-Bus protocol.

The prior implementation of the notifications code reserved a single
notification identifier when the program started; however, some
notification daemons (like mate-notification-daemon) automatically
terminate after some period of notification inactivity. They are
reactivated by writing new notifications via D-Bus. In this case, the
generated ID would no longer make sense to the newly activated process,
and show the notification would fail.

With the help of the new library, this change keeps track of the
notification ID and only reserves it while the notification is actually
being shown.