fix: don't use keyboxd for gpg detection
Now the new detection heuristic is to check if we open the private key
files. Since we're detecting only the keys that need a yubikey touch, we filter the
ones that have the shadowed-private-key attribute on the S-expression
inside the key file ( more information about it here:
https://github.com/gpg/gnupg/blob/6737e07a9b04064947ae37abd28b845a09abee22/doc/keyformat.txt#shadowed-private-key-format )

We can have false positives if the user has a yubikey + other gpg
smartcard device (since will detect it will try to open the private key
file too!).

This closes #54